Changed amount datatype
By switching to integer from decimal it doesn`t add .0 on re-display

+++ app/models/forms/application/savings_investment.rb
@@ -9,7 +9,7 @@ module Forms
           min_threshold_exceeded: Boolean,
           over_61: Boolean,
           max_threshold_exceeded: Boolean,
-          amount: Decimal
+          amount: Integer
         }
       end
